requirements.txt
django
djangorestframework
Pillow
drf-yasg
django-cors-headers
